+++ lib/src/services/settings_service.dart
@@ -38,4 +38,45 @@ class SettingsService extends BaseService {
         )
         .then((data) => data as Map<String, dynamic>? ?? {});
   }
+
+  /// Performs a S3 storage connection test.
+  Future<void> testS3({
+    Map<String, dynamic> body = const {},
+    Map<String, dynamic> query = const {},
+    Map<String, String> headers = const {},
+  }) {
+    return client.send(
+      "/api/settings/test/s3",
+      method: "POST",
+      body: body,
+      query: query,
+      headers: headers,
+    );
+  }
+
+  /// Sends a test email.
+  ///
+  /// The possible `template` values are:
+  /// - verification
+  /// - password-reset
+  /// - email-change
+  Future<void> testEmail(
+    String toEmail,
+    String template, {
+    Map<String, dynamic> body = const {},
+    Map<String, dynamic> query = const {},
+    Map<String, String> headers = const {},
+  }) {
+    final enrichedBody = Map<String, dynamic>.of(body);
+    enrichedBody["email"] ??= toEmail;
+    enrichedBody["template"] ??= template;
+
+    return client.send(
+      "/api/settings/test/email",
+      method: "POST",
+      body: enrichedBody,
+      query: query,
+      headers: headers,
+    );
+  }
 }